Additional logging for audio playback and also let playbacks finish
* Add additional logging to help profile audio playback. Playback was
stuttering on my computer, but it might not be a problem with dedicated
soundcards?

* Currently, we trigger playback and then wait for the stim duration for
before cutting off the stream. However, for some reason the timing isnt
quite right and can cut off the playbacks early. So I added some code
that will wait for the playbacks to finish (sometimes up to 0.5s!)
before the code moves on. This must be fixed...

* Some small changes to be forward compatible with Python 3